Update list of base packages, ignore package_version warnings
Add some missing packages according to the docs [1].

Also temporarily disable the package_version warnings because it's
complaining about a more recent version of docker that it is expecting.

Docker 1.13 is actually not blacklisted on 3.7 [2], we likely need to
rebase our openshift-ansible fork with our patches to get rid of the
warning -- and then we can remove the package_version exception.

[1]: https://docs.openshift.com/container-platform/3.7/install_config/install/host_preparation.html
[2]: https://github.com/openshift/origin/blob/5eda3fafdbc333c5a399f127e70b66aa324e3ffc/origin.spec#L404-L407

Change-Id: I7f8833b88a03c6a64a9b0b0565c441bdcc91c499